pulsewatch/requirements.txt
Flask==3.1.0
psutil==5.9.8
speedtest-cli==2.1.3